EXIF information is carried across resize
When images are resized using the PIL library the EXIF information is
lost. This fixes that by copying the information from the old picture to
new new resized one with the exception of the width and height figures
which are overwritten with the values of the new resized picture.

I suppose that the dpi (X/Y-resolution) figures should be updated too!?

Fix for #8.