Add Apache 2.0 license file.
I checked with Nathan Van Gheem, author of a Flicker-to-Picasa script that
I based this script on, to see what license his orginal script was using.

He wrote back:

I never put a license on it but you are free to use it however you want. I
don't care :)

After considering the pros and cons of different licenses, I chose the

Apache 2.0 license:

http://www.apache.org/licenses/LICENSE-2.0.txt